Move CAC/RAVE scripts to a sub-directory
