+++ libs/auth/scripts/cac/cac-get-cert
