+++ libs/auth/scripts/cac/configure-simulated-cac-card
