--- libs/auth/scripts/configure-dev-simulated-cac-card
@@ -1,9 +1,0 @@
-#!/usr/bin/env bash
-
-set -euo pipefail
-
-SCRIPTS_DIRECTORY="$(dirname "${BASH_SOURCE[0]}")"
-
-VX_CERT_AUTHORITY_CERT_PATH="${SCRIPTS_DIRECTORY}/../certs/dev/vx-cert-authority-cert.pem" \
-    VX_PRIVATE_KEY_PATH="${SCRIPTS_DIRECTORY}/../certs/dev/vx-private-key.pem" \
-    "${SCRIPTS_DIRECTORY}/configure-simulated-cac-card"
